+++ test/integration/shpec/ide_no_rm.sh
@@ -1,4 +1,4 @@
-describe "ide --rm false"
+describe "ide --no_rm"
   # make absolute path out of relative
   IDE_PATH=$(readlink -f "./ide")
   publicide_path="test/docker/publicide-usage"
@@ -7,19 +7,19 @@ describe "ide --rm false"
   publicide_path_dc="test/docker-compose/publicide-usage"
 
   describe 'when IDE_DRIVER="docker"'
-    describe 'when --dryrun and --rm false is set'
+    describe 'when --dryrun and --no_rm is set'
       it "returns 0; creates iderc, does not create docker container"
         rm -rf "${iderc}"
-        message="$(cd ${publicide_path} && ${IDE_PATH} --rm false --dryrun)"
+        message="$(cd ${publicide_path} && ${IDE_PATH} --no_rm --dryrun)"
         assert equal "$?" "0"
         assert do_not_match "$message" "--rm"
         rm -rf "${iderc}"
       end
     end
-    describe 'when --rm false is set'
+    describe 'when --no_rm is set'
       it "returns 0; creates docker container and does not remove it"
         rm -rf "${iderc}"
-        message="$(cd ${publicide_path} && IDE_LOG_LEVEL=debug ${IDE_PATH} --rm false whoami)"
+        message="$(cd ${publicide_path} && IDE_LOG_LEVEL=debug ${IDE_PATH} --no_rm whoami)"
         assert equal "$?" "0"
         assert do_not_match "$message" "--rm"
 
@@ -35,11 +35,11 @@ describe "ide --rm false"
     end
   end
   describe 'when IDE_DRIVER="docker-compose"'
-    describe 'when --dryrun and --rm false is set'
+    describe 'when --dryrun and --no_rm is set'
       it "returns 1; not implemented"
-        message="$(cd ${publicide_path_dc} && ${IDE_PATH} --rm false --dryrun)"
+        message="$(cd ${publicide_path_dc} && ${IDE_PATH} --no_rm --dryrun)"
         assert equal "$?" "1"
-        assert do_match "$message" "Not implemented feature: '--rm false' for driver: docker-compose"
+        assert do_match "$message" "Not implemented feature: '--no_rm' for driver: docker-compose"
       end
     end
   end